Move scraper scripts around and clean up Schamper & Urgent.fm
