--- scraper/.gitignore
@@ -1,10 +1,0 @@
-resto/1.0/menu
-resto/1.0/week
-resto/2.0/*/menu
-resto/2.0/menu
-resto/2.0/extrafood.json
-resto/2.0/sandwiches.json
-schamper
-urgentfm
-__pycache__/
-.venv/
