--- scraper/urgentfm.sh
@@ -1,6 +1,0 @@
-#!/usr/bin/env bash
-
-set -euo pipefail
-
-python3 urgentfm.py
-rsync -a urgentfm ~/public/api/2.0
