+++ server/scraper/resto/menu_manual.py
