+++ server/scraper/schamper.sh
