+++ server/scraper/util.py
@@ -1,5 +1,6 @@
 import re
 import sys
+import json
 
 
 def parse_money(moneystring):
@@ -8,3 +9,11 @@ def parse_money(moneystring):
 
 def stderr_print(*args, **kwargs):
     print(*args, file=sys.stderr, **kwargs)
+
+
+def write_json_to_file(obj, path):
+    """
+    Write an object to JSON at the specified path.
+    """
+    with open(path, mode='w') as f:
+        json.dump(obj, f, sort_keys=True)
